moving geowave-security project to inside of core module and deleting security module.
