+++ core/security/src/main/java/mil/nga/giat/geowave/security/utils/SecurityUtils.java
