+++ core/security/src/main/resources/geowave_crypto_key.dat
