+++ core/security/src/main/resources/log4j.properties
