+++ core/security/src/test/resources/geowave_crypto_key.dat
