+++ core/security/src/test/resources/geowave_crypto_key2.dat
