+++ pom.xml
@@ -737,10 +737,6 @@
 									<title>Services</title>
 									<packages>mil.nga.giat.geowave.service*</packages>
 								</group>
-                                <group>
-                                    <title>Security</title>
-                                    <packages>mil.nga.giat.geowave.security*</packages>
-                                </group>
 								<group>
 									<title>Examples</title>
 									<packages>mil.nga.giat.geowave.example*</packages>
@@ -816,7 +812,6 @@
 		<module>examples</module>
 		<module>extensions</module>
 		<module>services</module>
-        <module>security</module>
 		<module>test</module>
 	</modules>
 	<repositories>
