--- security/pom.xml
@@ -1,16 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<artifactId>geowave-parent</artifactId>
-		<groupId>mil.nga.giat</groupId>
-		<relativePath>../</relativePath>
-		<version>0.9.4-SNAPSHOT</version>
-	</parent>
-	<artifactId>geowave-security-parent</artifactId>
-	<name>GeoWave Security Parent POM</name>
-	<description>Security options &amp; implementations provided for GeoWave</description>
-	<packaging>pom</packaging>
-	<modules>
-		<module>geowave-security</module>
-	</modules>
-</project>
